Updated exercise 4
This commit updates exercise 4. Parts where solutions are found from R demos are removed, and
parts about computing MCSEs are added. Rubric is also updated. No longer needed functions
and markmyassignment tests are removed.

This commit also adds the mcse_quantile function to the aaltobda package.
This function is used in the exercise to compute MCSE of quantiles for independent draws.
The function is a simplified version of a similar function from the loo package,
but now instead of effective samples size, the actual sample size is used.